third_party/WebKit/LayoutTests/fast/css/css-escaped-identifier.html

<html>
<head>
    <!-- The testcase input -->
    <style>
        d\eeeex {}
        D\eeeeX {}
    </style>

    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>

    <script>
        test(function() {
            var styleElm = document.getElementsByTagName("style")[0];
            assert_not_equals(styleElm, null, "Style element exists");
            var stylesheet = styleElm.sheet;
            assert_not_equals(stylesheet, undefined, "style.sheet exists");
            rules = stylesheet.cssRules;
            assert_equals(rules.length, 2, "cssRules.length == 2");
            assert_equals(rules[0].selectorText, "d\ueeeex", "Lowercase identifier with escapes");
            assert_equals(rules[1].selectorText, "d\ueeeex", "Uppercase identifier with escapes");
        }, "Correctly parsed identifier with escapes");
    </script>
</head>
<body><!-- Intentionally left empty --></body>
</html>
